Add spec to try and make COUNTBLANK work
Currently we 'fix' values by converting nils and empty strings to 0.0 but this isn't ideally what we would do.

Ideally we'd allow the user to handle these in the spreadsheet and apply that handling corrently. Where the grammar currently falls down is that it allows strings to be included in numerical expressions. ie expression, multiplicative, etc don't require that the arguments are numerical. If the grammar supported this when it compiled to ruby we could maintain the values internally and JIT convert them depending on the type of expression.